+++ lib/clang-utils.js
